Renamed kg-default-nodes/test/utils to test-utils
no issue

- reduces ambiguity/confusion when adding tests for the src package utils files
