+++ packages/kg-default-nodes/test/nodes/at-link.test.js
@@ -1,4 +1,4 @@
-require('../utils');
+require('../test-utils');
 const {JSDOM} = require('jsdom');
 const {createHeadlessEditor} = require('@lexical/headless');
 const {AtLinkNode, $createAtLinkNode, $isAtLinkNode, $createAtLinkSearchNode, AtLinkSearchNode} = require('../../');
