+++ packages/kg-default-nodes/test/nodes/codeblock.test.js
@@ -1,4 +1,4 @@
-const {createDocument, dom, html} = require('../utils');
+const {createDocument, dom, html} = require('../test-utils');
 const {createHeadlessEditor} = require('@lexical/headless');
 const {$generateNodesFromDOM} = require('@lexical/html');
 const {CodeBlockNode, $createCodeBlockNode, $isCodeBlockNode} = require('../../');
