+++ packages/kg-default-nodes/test/nodes/collection.test.js
@@ -1,4 +1,4 @@
-const {createDocument, dom, html} = require('../utils');
+const {createDocument, dom, html} = require('../test-utils');
 const {createHeadlessEditor} = require('@lexical/headless');
 const {$generateNodesFromDOM} = require('@lexical/html');
 const {CollectionNode, $createCollectionNode, $isCollectionNode} = require('../../');
